Move ReactFlightClientConfigBundlerParcel to the client folder
Aligns with the others.
