+++ packages/react-server-dom-parcel/src/client/ReactFlightDOMClientBrowser.js
@@ -10,7 +10,7 @@
 import type {Thenable} from 'shared/ReactTypes.js';
 import type {Response as FlightResponse} from 'react-client/src/ReactFlightClient';
 import type {ReactServerValue} from 'react-client/src/ReactFlightReplyClient';
-import type {ServerReferenceId} from '../shared/ReactFlightClientConfigBundlerParcel';
+import type {ServerReferenceId} from '../client/ReactFlightClientConfigBundlerParcel';
 
 import {
   createResponse,
